Refine akhil-square-shadow overlay and cleanup UI comments
