+++ src/themes/akhil-square-shadow/_landing-page.scss
@@ -57,3 +57,69 @@
     opacity: 0.4;
   }
 }
+/**
+ * Stars - Twinkling starfield for night sky
+ *
+ * Container for randomly positioned stars that fade in during night transitions.
+ * Light mode: opacity 0 (hidden)
+ * Night mode: opacity transitions to visible
+ */
+.landing-stars {
+  position: absolute;
+  inset: 0;
+  opacity: 0;
+  transition: opacity 1s ease-out;
+  pointer-events: none;
+}
+
+/**
+ * Individual star element.
+ *
+ * Light/dark mode variant support:
+ * - Light mode: white stars on day background
+ * - Dark mode: white stars on night background (always visible)
+ *
+ * Uses random sizing and positioning to create natural starfield.
+ */
+.landing-star {
+  position: absolute;
+  background: white;
+  border-radius: 50%;
+}
+
+/**
+ * Sun - Day circle with golden glow
+ *
+ * Radial gradient creates depth with bright center and darker edges.
+ * Glow effect uses layered box-shadows.
+ * Positioned with transform: translate(-50%, -50%) for arc positioning.
+ */
+.landing-sun {
+  position: absolute;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  background: radial-gradient(circle at 30% 30%, #fff5d4 0%, #ffd700 40%, #ff8c00 100%);
+  box-shadow: 0 0 30px #ffd700, 0 0 60px rgba(255, 215, 0, 0.4);
+  transform: translate(-50%, -50%);
+  transition: opacity 0.5s ease-out;
+}
+
+/**
+ * Moon - Night circle with subtle gray gradient
+ *
+ * Subtle radial gradient creates dimensionality.
+ * Soft glow effect with lower opacity shadow.
+ * Positioned with transform: translate(-50%, -50%) for arc positioning.
+ */
+.landing-moon {
+  position: absolute;
+  width: 30px;
+  height: 30px;
+  border-radius: 50%;
+  background: radial-gradient(circle at 40% 40%, #f5f5f5 0%, #dcdcdc 50%, #a9a9a9 100%);
+  box-shadow: 0 0 20px rgba(255, 255, 255, 0.4);
+  transform: translate(-50%, -50%);
+  opacity: 0;
+  transition: opacity 0.5s ease-out;
+}
